# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Document Generator API
  description: |
    API de génération de PDF par lots : file d’attente asynchrone (**Bull** sur Redis, ou file en mémoire
    `QUEUE_BACKEND=memory`), stockage binaire **GridFS** (MongoDB), génération CPU dans des **worker threads**,
    disjoncteurs **Opossum** (GridFS, DocuSign simulé), logs **JSON** (corrélation `x-correlation-id`, `batchId`, `documentId`)
    et métriques **Prometheus**.

    **Swagger UI** : `/api-docs` — spécification source : `docs/openapi.yaml`.

    **Tests de l’API** : collection Postman `postman/document-generator.collection.json` (`baseUrl`, `workerMetricsUrl`).

    **Courbes de benchmark** : à la racine du dépôt, `benchmark-series-excel-fr-avecgraph.xlsx` (graphiques sur séries de charge).
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Développement local
...
paths:
...
  /batch:
    post:
      tags: [documents]
      summary: Créer un lot de documents PDF
      operationId: createBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBatchRequest'
            examples:
              sample:
                summary: Un document
                value:
                  documents:
                    - title: Rapport Q1
                      content: "Contenu du rapport avec données et conclusion."
              multi:
                summary: Plusieurs documents
                value:
                  documents:
                    - title: Note A
                      content: "Texte court."
                    - title: Note B
                      content: "Autre contenu (max 200k caractères par document)."
      responses:
        '202':
          description: Lot accepté pour traitement asynchrone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBatchResponse'
              examples:
                sample:
                  summary: Réponse type
                  value:
                    batchId: 65f0c4e2b8a1c2d3e4f5a6b7
                    documentIds:
                      - 65f0c4e2b8a1c2d3e4f5a6b8
                multi:
                  summary: Plusieurs IDs créés
                  value:
                    batchId: 65f0c4e2b8a1c2d3e4f5a6b7
                    documentIds:
                      - 65f0c4e2b8a1c2d3e4f5a6b8
                      - 65f0c4e2b8a1c2d3e4f5a6b9
                      - 65f0c4e2b8a1c2d3e4f5a6ba
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /batch/{id}:
    get:
      tags: [documents]
      summary: Consulter l’état d’un lot
      operationId: getBatch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-fA-F0-9]{24}$'
          example: "65f0c4e2b8a1c2d3e4f5a6b7"
      responses:
        '200':
          description: Détails du lot et des documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchDetails'
              examples:
                partial_failure:
                  summary: Lot partiel (certains PDF en échec)
                  value:
                    batchId: 65f0c4e2b8a1c2d3e4f5a6b7
                    status: partial
                    createdAt: '2025-03-27T10:00:00.000Z'
                    updatedAt: '2025-03-27T10:05:00.000Z'
                    documents:
                      - documentId: 65f0c4e2b8a1c2d3e4f5a6b8
                        title: OK
                        status: completed
                        gridFsFileId: 65f0c4e2b8a1c2d3e4f5a6c0
                      - documentId: 65f0c4e2b8a1c2d3e4f5a6b9
                        title: KO
                        status: failed
                        errorMessage: "Génération PDF : délai de 5000 ms dépassé"
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
...
    CreateBatchRequest:
      type: object
      example:
        documents:
          - title: Contrat type
            content: "Les parties conviennent des conditions suivantes…"
      required: [documents]
      properties:
        documents:
          type: array
          minItems: 1
          maxItems: 1000
          description: Taille max configurable côté serveur (défaut 1000, variable BATCH_MAX_DOCUMENTS, plafond 5000).
          items:
            type: object
            required: [title, content]
            properties:
              title:
                type: string
                maxLength: 500
              content:
                type: string
                maxLength: 200000
    CreateBatchResponse:
      type: object
      required: [batchId, documentIds]
      properties:
        batchId:
          type: string
          example: "65f0c4e2b8a1c2d3e4f5a6b7"
        documentIds:
          type: array
          items:
            type: string
    DocumentStatusItem:
      type: object
      properties:
        documentId:
          type: string
        title:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed]
        gridFsFileId:
          type: string
          nullable: true
        errorMessage:
          type: string
          nullable: true
    BatchDetails:
      type: object
      required: [batchId, status, createdAt, updatedAt, documents]
      properties:
        batchId:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed, partial]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentStatusItem'
...
    ErrorBody:
      type: object
      example:
        error:
          message: Description lisible de l’erreur
          code: VALIDATION_ERROR
      properties:
        error:
          type: object
          properties:
            message:
              type: string
              example: Array must contain at least 1 element(s)
            code:
              type: string
              example: VALIDATION_ERROR
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            error:
              message: "Array must contain at least 1 element(s)"
              code: VALIDATION_ERROR
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            error:
              message: "Lot introuvable"
              code: NOT_FOUND
...
    TooManyRequests:
      description: Limite de débit
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            error:
              message: "Trop de requêtes"
              code: RATE_LIMIT
    InternalError:
      description: Erreur serveur
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            error:
              message: "Erreur interne du serveur"
              code: INTERNAL_ERROR